requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
httpx>=0.28.0
twilio>=9.0.0
openai>=1.60.0
python-dotenv>=1.0.0
